Upgrade to PY from edge branch for Streamlink 3 compatibility
Check and update the Python version on the container and upgrade
it to the latest version on Alpine's edge branch whenever it falls
behind.  It also installs/upgrades other related Python packages
built for the target Alpine branch and architecture. This avoids
issues with trying to build Streamlink requirements from source (see
issue #6).

Improve readability and other minor changes to the script.